refactor: move storage file handling to reusable composable
